% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{addin_join_inspector}
\alias{addin_join_inspector}
\title{'RStudio' Addin: Join Inspector}
\usage{
addin_join_inspector()
}
\value{
Called for side effects (opens Shiny gadget).
}
\description{
Opens an interactive dialog to explore join diagnostics between two
data frames in the current R environment.
}
\keyword{internal}
