% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_spy.R
\name{.format_bytes}
\alias{.format_bytes}
\title{Format Bytes as Human-Readable String}
\usage{
.format_bytes(bytes)
}
\arguments{
\item{bytes}{Numeric bytes.}
}
\value{
Character string.
}
\description{
Format Bytes as Human-Readable String
}
\keyword{internal}
