% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_diagnostics.R
\name{.levenshtein}
\alias{.levenshtein}
\title{Calculate 'Levenshtein' Distance}
\usage{
.levenshtein(s1, s2)
}
\arguments{
\item{s1}{First string.}

\item{s2}{Second string.}
}
\value{
Integer distance.
}
\description{
Simple implementation for detecting near-matches.
}
\keyword{internal}
