% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.report_to_list}
\alias{.report_to_list}
\title{Convert JoinReport to list for serialization}
\usage{
.report_to_list(report, timestamp = TRUE)
}
\arguments{
\item{report}{A JoinReport object.}

\item{timestamp}{Include timestamp.}
}
\value{
List representation.
}
\description{
Convert JoinReport to list for serialization
}
\keyword{internal}
