% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.report_to_text}
\alias{.report_to_text}
\title{Convert JoinReport to text}
\usage{
.report_to_text(report, timestamp = TRUE)
}
\arguments{
\item{report}{A JoinReport object.}

\item{timestamp}{Include timestamp.}
}
\value{
Character string.
}
\description{
Convert JoinReport to text
}
\keyword{internal}
