% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.to_json}
\alias{.to_json}
\title{Simple JSON serialization (no dependencies)}
\usage{
.to_json(x)
}
\arguments{
\item{x}{List to convert.}
}
\value{
JSON string.
}
\description{
Simple JSON serialization (no dependencies)
}
\keyword{internal}
