% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-keys.R
\name{.validate_df}
\alias{.validate_df}
\title{Validate that an object is a data frame}
\usage{
.validate_df(x, arg, call = rlang::caller_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name used in the error message.}

\item{call}{Environment to attribute the error to.}
}
\value{
Invisibly, \code{x}.
}
\description{
Validate that an object is a data frame
}
\keyword{internal}
