% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_log_file}
\alias{get_log_file}
\title{Get Current Log File}
\usage{
get_log_file()
}
\value{
The log file path, or \code{NULL} if not set.
}
\description{
Returns the current automatic log file path, if set.
}
\seealso{
\code{\link[=set_log_file]{set_log_file()}}
}
