% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinReport.R
\name{is_join_report}
\alias{is_join_report}
\title{Check if Object is a JoinReport}
\usage{
is_join_report(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{JoinReport}, \code{FALSE} otherwise.
}
\description{
Check if Object is a JoinReport
}
