% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinReport.R
\name{new_join_report}
\alias{new_join_report}
\title{Create a JoinReport Object}
\usage{
new_join_report(
  x_summary,
  y_summary,
  match_analysis,
  issues,
  expected_rows,
  by
)
}
\arguments{
\item{x_summary}{Summary statistics for keys in the left table.}

\item{y_summary}{Summary statistics for keys in the right table.}

\item{match_analysis}{Details of which keys will/won't match.}

\item{issues}{List of detected problems.}

\item{expected_rows}{Predicted row counts for each join type.}

\item{by}{The columns used for joining.}
}
\value{
A \code{JoinReport} S3 object.
}
\description{
Create a JoinReport Object
}
\keyword{internal}
