% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinReport.R
\name{print.JoinReport}
\alias{print.JoinReport}
\title{Print Method for JoinReport}
\usage{
\method{print}{JoinReport}(x, ...)
}
\arguments{
\item{x}{A \code{JoinReport} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for JoinReport
}
